refactor(api-1d): move school files related to assessment & activity
