+++ api/src/school/domain/usecases/get-current-activity.js
