+++ api/src/shared/application/assessments/assessment-controller.js
@@ -8,7 +8,6 @@ import * as events from '../../../../lib/domain/events/index.js';
 import { logger } from '../../../../lib/infrastructure/logger.js';
 import * as assessmentRepository from '../../infrastructure/repositories/assessment-repository.js';
 import * as assessmentSerializer from '../../infrastructure/serializers/jsonapi/assessment-serializer.js';
-import * as activitySerializer from '../../../../lib/infrastructure/serializers/jsonapi/activity-serializer.js';
 import * as challengeSerializer from '../../infrastructure/serializers/jsonapi/challenge-serializer.js';
 import * as competenceEvaluationSerializer from '../../../evaluation/infrastructure/serializers/jsonapi/competence-evaluation-serializer.js';
 import {
@@ -87,12 +86,6 @@ const getNextChallenge = async function (
   }
 };
 
-const getCurrentActivity = async function (request, h, dependencies = { activitySerializer }) {
-  const assessmentId = request.params.id;
-  const activity = await usecases.getCurrentActivity({ assessmentId });
-  return dependencies.activitySerializer.serialize(activity);
-};
-
 const completeAssessment = async function (request) {
   const assessmentId = request.params.id;
   const locale = extractLocaleFromRequest(request);
@@ -165,7 +158,6 @@ const assessmentController = {
   get,
   getLastChallengeId,
   getNextChallenge,
-  getCurrentActivity,
   completeAssessment,
   updateLastChallengeState,
   findCompetenceEvaluations,
