+++ api/src/shared/application/assessments/index.js
@@ -45,22 +45,6 @@ const register = async function (server) {
     },
     {
       method: 'GET',
-      path: '/api/pix1d/assessments/{id}/current-activity',
-      config: {
-        pre: [{ method: securityPreHandlers.checkPix1dActivated }],
-        auth: false,
-        validate: {
-          params: Joi.object({
-            id: identifiersType.assessmentId,
-          }),
-        },
-        handler: assessmentController.getCurrentActivity,
-        notes: ["- Récupération de l'activité courante"],
-        tags: ['api'],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/assessments/{id}',
       config: {
         auth: false,
