+++ api/tests/school/unit/application/assessment-controller_test.js
@@ -1,6 +1,7 @@
 import { expect, hFake, sinon } from '../../../test-helper.js';
 import { assessmentController } from '../../../../src/school/application/assessment-controller.js';
 import { usecases } from '../../../../src/school/domain/usecases/index.js';
+import { Activity } from '../../../../src/school/domain/models/Activity.js';
 
 describe('Unit | Controller | assessment-controller', function () {
   describe('#getNextChallengeForPix1d', function () {
@@ -26,6 +27,36 @@ describe('Unit | Controller | assessment-controller', function () {
 
       // then
       expect(result).to.be.equal(challenge);
+    });
+  });
+
+  describe('#getCurrentActivity', function () {
+    let activity;
+    const assessmentId = 104974;
+    let activitySerializerStub;
+
+    beforeEach(function () {
+      activity = { assessmentId, level: Activity.levels.TUTORIAL };
+      sinon.stub(usecases, 'getCurrentActivity').withArgs({ assessmentId }).resolves(activity);
+      activitySerializerStub = { serialize: sinon.stub() };
+      activitySerializerStub.serialize.resolvesArg(0);
+    });
+
+    it('should call the expected usecase', async function () {
+      // given
+      const request = {
+        params: {
+          id: assessmentId,
+        },
+      };
+
+      // when
+      const result = await assessmentController.getCurrentActivity(request, hFake, {
+        activitySerializer: activitySerializerStub,
+      });
+
+      // then
+      expect(result).to.be.equal(activity);
     });
   });
 });
