+++ api/tests/shared/unit/application/assessments/assessment-controller_test.js
@@ -6,7 +6,6 @@ import { usecases as certificationUsecases } from '../../../../../src/certificat
 import * as events from '../../../../../lib/domain/events/index.js';
 import { AssessmentCompleted } from '../../../../../lib/domain/events/AssessmentCompleted.js';
 import { DomainTransaction } from '../../../../../lib/infrastructure/DomainTransaction.js';
-import { Activity } from '../../../../../src/school/domain/models/Activity.js';
 
 describe('Unit | Controller | assessment-controller', function () {
   describe('#createAssessmentPreviewForPix1d', function () {
@@ -57,36 +56,6 @@ describe('Unit | Controller | assessment-controller', function () {
 
       // then
       expect(result).to.be.equal(assessment);
-    });
-  });
-
-  describe('#getCurrentActivity', function () {
-    let activity;
-    const assessmentId = 104974;
-    let activitySerializerStub;
-
-    beforeEach(function () {
-      activity = { assessmentId, level: Activity.levels.TUTORIAL };
-      sinon.stub(usecases, 'getCurrentActivity').withArgs({ assessmentId }).resolves(activity);
-      activitySerializerStub = { serialize: sinon.stub() };
-      activitySerializerStub.serialize.resolvesArg(0);
-    });
-
-    it('should call the expected usecase', async function () {
-      // given
-      const request = {
-        params: {
-          id: assessmentId,
-        },
-      };
-
-      // when
-      const result = await assessmentController.getCurrentActivity(request, hFake, {
-        activitySerializer: activitySerializerStub,
-      });
-
-      // then
-      expect(result).to.be.equal(activity);
     });
   });
